# modules/set_metadata: add shift and mask support