This patch changes SetMetadata to optionally shift and mask (in that
order) values when copying metadata values from the contents of a
packet. It also fixes one of the metadata test scripts.

A quick micro benchmark on a 2.1GHz Xeon 2620v4 with the script below shows relatively minor overhead is imposed when `SetMetadata` is not configured to use the new features.

```
ATTR_SIZE = int($ATTR_SIZE!'1')
Source() \
    -> SetMetadata(attrs=[{'name': 'foo', 'size': ATTR_SIZE, 'offset': 0}]) \
    -> MetadataTest(read={'foo': ATTR_SIZE}) \
    -> Sink()
```

### Before

| attr. size | Mpps | cycles/pkt |
| - | - | - |
| 1 | 108.108 | 19.422 |
| 2 | 108.407 | 19.372 |
| 4 | 111.280 | 18.872 |
| 8 | 113.383 | 18.522 |
| 16 | 106.242 | 19.767 |
| 24 | 115.461 | 18.188 |
| 32 | 116.844 | 17.973 |

### After

|attr. size | Mpps | cycles/pkt |
| - | - | - |
| 1 | 103.293 | 20.331 |
| 2 | 103.518 | 20.287 |
| 4 | 109.368 | 19.202 |
| 8 | 109.321 | 19.210 |
| 16 | 109.904 | 19.108 |
| 24 | 111.408 | 18.850 |
| 32 | 114.985 | 18.264 |