@melvinw 2017-08-02
Thanks for the comments! I'll see if I can come up with a way to use `_mm_bslli_si128()` and `_mm_bsrli_si128()` to handle the shifting, though it's not immediately clear how I'd do that with a carry (maybe I don't need it).