@melvinw 2017-08-03
Going to reopen this as a slightly different PR